# docs/api/admin.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "admin-service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/audit/log": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "List Audit Log",
        "description": "Paginated ``audit_log`` listing (newest first, cursor by ``id``).",
        "operationId": "list_audit_log_admin_audit_log_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return rows with id less than this value (older page).",
              "title": "Cursor"
            },
            "description": "Return rows with id less than this value (older page)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/audit/verify": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Verify Audit Log",
        "description": "Verify hash-chain integrity for a timestamp range via audit-service.",
        "operationId": "verify_audit_log_admin_audit_verify_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/backtest": {
      "get": {
        "tags": [
          "backtest"
        ],
        "summary": "List Recent Backtests",
        "description": "List recent backtest runs (newest first).",
        "operationId": "list_recent_backtests_admin_backtest_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "backtest"
        ],
        "summary": "Start Backtest",
        "description": "Forward to ``backtest-engine POST /backtest/run`` and audit log.",
        "operationId": "start_backtest_admin_backtest_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartBacktestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartBacktestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/backtest/{backtest_id}/results": {
      "get": {
        "tags": [
          "backtest"
        ],
        "summary": "Get Backtest Results",
        "description": "Proxy ``backtest-engine GET /backtest/{id}/results``.",
        "operationId": "get_backtest_results_admin_backtest__backtest_id__results_get",
        "parameters": [
          {
            "name": "backtest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Backtest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResultsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditLogEntry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "ts",
          "actor",
          "action",
          "entity_type",
          "entity_id"
        ],
        "title": "AuditLogEntry",
        "description": "One row from ``audit_log``."
      },
      "AuditLogPageResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AuditLogEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass as ``cursor`` query param to fetch the next page (older rows)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entries",
          "limit"
        ],
        "title": "AuditLogPageResponse",
        "description": "``GET /admin/audit/log`` paginated result."
      },
      "AuditVerifyResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "mismatch_at_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mismatch At Id"
          },
          "rows_checked": {
            "type": "integer",
            "title": "Rows Checked",
            "default": 0
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "AuditVerifyResponse",
        "description": "``GET /admin/audit/verify`` \u2014 hash-chain check via audit-service."
      },
      "BacktestListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/BacktestRunSummary"
            },
            "type": "array",
            "title": "Runs"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runs",
          "limit"
        ],
        "title": "BacktestListResponse",
        "description": "``GET /admin/backtest`` payload."
      },
      "BacktestResultsResponse": {
        "properties": {
          "backtest_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Backtest Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "result_blob_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Blob Uri"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "backtest_run_id",
          "status"
        ],
        "title": "BacktestResultsResponse",
        "description": "Proxied response from backtest-engine ``GET /backtest/{id}/results``."
      },
      "BacktestRunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "universe": {
            "type": "string",
            "title": "Universe"
          },
          "git_sha": {
            "type": "string",
            "title": "Git Sha"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result_blob_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Blob Uri"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "strategy_id",
          "start_date",
          "end_date",
          "universe",
          "git_sha",
          "started_at",
          "status"
        ],
        "title": "BacktestRunSummary",
        "description": "One row from ``theeyebeta.backtest_runs`` for the listing."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StartBacktestRequest": {
        "properties": {
          "strategy_id": {
            "type": "string",
            "minLength": 1,
            "title": "Strategy Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "universe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Universe"
          },
          "walk_forward": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Walk Forward"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(replay|redecision)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "strategy_id",
          "start_date",
          "end_date"
        ],
        "title": "StartBacktestRequest",
        "description": "``POST /admin/backtest`` body \u2014 forwarded to backtest-engine."
      },
      "StartBacktestResponse": {
        "properties": {
          "backtest_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Backtest Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "backtest_run_id",
          "status"
        ],
        "title": "StartBacktestResponse",
        "description": "Proxied response from backtest-engine ``POST /backtest/run``."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}